+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * [TT-4433] Add network interceptor for CRSF token
 * [TT-4429] Fix routes product_type_id parameter
 * [TT-4432] Add endpoints for updating and retrieving bookings
+* [TT-4435] Add support for checkout endpoint
 
 ## 0.5.0